requirements.txt
PyQt6
uiautomator2
adbutils
requests
Pillow
